+++ src/main/kotlin/com/rtomyj/skc/model/Source.kt
@@ -1,4 +1,4 @@
-package com.rtomyj.skc.skcsuggestionengine.traffic.model
+package com.rtomyj.skc.model
 
 data class Source(
 	val systemName: String,
