+++ src/main/kotlin/com/rtomyj/skc/model/SuggestionEngineStatus.kt
@@ -1,4 +1,4 @@
-package com.rtomyj.skc.skcsuggestionengine.status.model
+package com.rtomyj.skc.model
 
 data class SuggestionEngineStatus(
 	val version: String,
