+++ src/main/kotlin/com/rtomyj/skc/model/Traffic.kt
@@ -1,0 +1,7 @@
+package com.rtomyj.skc.model
+
+data class Traffic(
+    val ip: String,
+    val source: Source,
+    val resourceUtilized: ResourceUtilized
+)
