--- src/main/kotlin/com/rtomyj/skc/skcsuggestionengine/traffic/model/Traffic.kt
@@ -1,7 +1,0 @@
-package com.rtomyj.skc.skcsuggestionengine.traffic.model
-
-data class Traffic(
-	val ip: String,
-	val source: Source,
-	val resourceUtilized: ResourceUtilized
-)
